Renaming dir go -> core
