+++ core/model_pin_meta.go
