+++ core/model_pin_results.go
