+++ core/model_pin_status.go
